Add support for requesting just the _id and _type of each hit.
Currently, you can only specify at least one field to be
returned, or the whole document.  There is no way to request
that no fields be returned.

WARNING: This is a breaking change. search.fields() now
expects a list of field names or None as an argument, instead
of *args where each argument is a field name.

Closes issue #26